chore: move safe call in an internal package
