+++ notice_test.go
@@ -124,57 +124,6 @@ func TestCreateWrappedFunction(t *testing.T) {
 	assert.Contains(t, loggerHandler.messages.String(), "[DEBUG] Nice this function returns cheese")
 }
 
-func TestSafeCall(t *testing.T) {
-	// Test a function that returns a string.
-	fn := func() (string, error) { return "cheese", nil }
-	out, err := safeCall(fn)
-	assert.NoError(t, err)
-	assert.Equal(t, "cheese", out)
-
-	// Test a function that returns a string and an error.
-	fn2 := func() (string, error) { return "cheese", fmt.Errorf("oh no") }
-	out, err = safeCall(fn2)
-	assert.Error(t, err)
-	assert.Equal(t, "cheese", out)
-
-	// Test a function that returns a string and an error.
-	fn3 := func() (string, error) { return "", fmt.Errorf("oh no") }
-	out, err = safeCall(fn3)
-	assert.Error(t, err)
-	assert.Empty(t, out)
-
-	// Test a function that returns a string and an error.
-	fn4 := func() (string, error) { return "", nil }
-	out, err = safeCall(fn4)
-	assert.NoError(t, err)
-	assert.Empty(t, out)
-
-	// Test a function that returns nothing.
-	fn5 := func() {}
-	out, err = safeCall(fn5)
-	assert.NoError(t, err)
-	assert.Nil(t, out)
-
-	// Test a function that returns 3 values.
-	a, b, c := "a", "b", "c"
-	fn6 := func(a, b, c string) (string, string, string) { return a, b, c }
-	out, err = safeCall(fn6, a, b, c)
-	assert.NoError(t, err)
-	assert.Equal(t, out, a, "the return should be the first argument")
-
-	// Test a case where the function panics.
-	fn7 := func() { panic("oh no") }
-	out, err = safeCall(fn7)
-	assert.ErrorContains(t, err, "recovered from panic: oh no")
-	assert.Nil(t, out)
-
-	// Test when fn is not a function.
-	fn8 := "cheese"
-	out, err = safeCall(fn8)
-	assert.ErrorContains(t, err, "fn is not a function")
-	assert.Nil(t, out)
-}
-
 func TestNoticeInTemplate(t *testing.T) {
 	loggerHandler := &noticeLoggerHandler{}
 	handler := New(WithLogger(slog.New(loggerHandler)))
